thermal-core: fix invalid sensor data being accepted
In some cases the sensor is supplying invalid data by
reporting the temperature as 0.

In this case we need to fallback to the DEFAULT_TEMP
which contains a safe value. This is better than just
relying on the last value supplied by the sensor since
it may be outdated at this time

This fixes the invalid behaviour of temperature
assisted scheduling based on temperature
sensor sampling in msm-core which led to this
issue being noticed.

Signed-off-by: Alex Naidis <alex.naidis@linux.com>